+++ mobile/android/android-components/components/feature/awesomebar/src/main/java/mozilla/components/feature/awesomebar/provider/HistoryMetadataSuggestionProvider.kt
@@ -15,28 +15,31 @@ import mozilla.components.concept.storage.HistoryStorage
 import mozilla.components.feature.session.SessionUseCases
 import java.util.UUID
 
-internal const val METADATA_SUGGESTION_LIMIT = 5
+/**
+ * Return 5 history suggestions by default.
+ */
+const val DEFAULT_METADATA_SUGGESTION_LIMIT = 5
 
 /**
  * A [AwesomeBar.SuggestionProvider] implementation that provides suggestions based on [HistoryMetadata].
  *
- * @property historyStorage an instance of the [HistoryStorage] used
+ * @param historyStorage an instance of the [HistoryStorage] used
  * to query matching metadata records.
- * @property loadUrlUseCase the use case invoked to load the url when the
+ * @param loadUrlUseCase the use case invoked to load the url when the
  * user clicks on the suggestion.
- * @property icons optional instance of [BrowserIcons] to load fav icons
+ * @param icons optional instance of [BrowserIcons] to load fav icons
  * for [HistoryMetadata] URLs.
  * @param engine optional [Engine] instance to call [Engine.speculativeConnect] for the
  * highest scored suggestion URL.
- * @param maxNumberOfResults optional parameter allowing to lower the number of returned suggested
- * history items to below the default of 5
+ * @param maxNumberOfSuggestions optional parameter to specify the maximum number of returned suggestions,
+ * defaults to [DEFAULT_METADATA_SUGGESTION_LIMIT].
  */
 class HistoryMetadataSuggestionProvider(
     private val historyStorage: HistoryMetadataStorage,
     private val loadUrlUseCase: SessionUseCases.LoadUrlUseCase,
     private val icons: BrowserIcons? = null,
     internal val engine: Engine? = null,
-    @VisibleForTesting internal val maxNumberOfResults: Int = -1
+    @VisibleForTesting internal val maxNumberOfSuggestions: Int = DEFAULT_METADATA_SUGGESTION_LIMIT
 ) : AwesomeBar.SuggestionProvider {
     override val id: String = UUID.randomUUID().toString()
 
@@ -48,18 +51,9 @@ class HistoryMetadataSuggestionProvider(
             return emptyList()
         }
 
-        // Having both maxNumberOfResults and METADATA_SUGGESTION_LIMIT ensures that when asking for
-        // a lower number of suggestions the below filtering will have some that can be dropped.
-        val maxReturnedSuggestions = if (maxNumberOfResults > 0) {
-            minOf(maxNumberOfResults, METADATA_SUGGESTION_LIMIT)
-        } else {
-            METADATA_SUGGESTION_LIMIT
-        }
-
         val suggestions = historyStorage
-            .queryHistoryMetadata(text, METADATA_SUGGESTION_LIMIT)
+            .queryHistoryMetadata(text, maxNumberOfSuggestions)
             .filter { it.totalViewTime > 0 }
-            .take(maxReturnedSuggestions)
 
         suggestions.firstOrNull()?.key?.url?.let { url -> engine?.speculativeConnect(url) }
         return suggestions.into()