+++ mobile/android/android-components/components/feature/awesomebar/src/main/java/mozilla/components/feature/awesomebar/provider/HistoryStorageSuggestionProvider.kt
@@ -15,29 +15,32 @@ import mozilla.components.feature.awesomebar.facts.emitHistorySuggestionClickedF
 import mozilla.components.feature.session.SessionUseCases
 import java.util.UUID
 
-internal const val HISTORY_SUGGESTION_LIMIT = 20
+/**
+ * Return 20 history suggestions by default.
+ */
+const val DEFAULT_HISTORY_SUGGESTION_LIMIT = 20
 
 /**
  * A [AwesomeBar.SuggestionProvider] implementation that provides suggestions based on the browsing
  * history stored in the [HistoryStorage].
  *
- * @property historyStorage and instance of the [HistoryStorage] used
+ * @param historyStorage and instance of the [HistoryStorage] used
  * to query matching history records.
- * @property loadUrlUseCase the use case invoked to load the url when the
+ * @param loadUrlUseCase the use case invoked to load the url when the
  * user clicks on the suggestion.
- * @property icons optional instance of [BrowserIcons] to load fav icons
+ * @param icons optional instance of [BrowserIcons] to load fav icons
  * for history URLs.
  * @param engine optional [Engine] instance to call [Engine.speculativeConnect] for the
  * highest scored suggestion URL.
- * @param maxNumberOfResults optional parameter allowing to lower the number of returned suggested
- * history items to below the default of 20
+ * @param maxNumberOfSuggestions optional parameter to specify the maximum number of returned suggestions,
+ * defaults to [DEFAULT_HISTORY_SUGGESTION_LIMIT]
  */
 class HistoryStorageSuggestionProvider(
     private val historyStorage: HistoryStorage,
     private val loadUrlUseCase: SessionUseCases.LoadUrlUseCase,
     private val icons: BrowserIcons? = null,
     internal val engine: Engine? = null,
-    @VisibleForTesting internal val maxNumberOfResults: Int = -1
+    @VisibleForTesting internal val maxNumberOfSuggestions: Int = DEFAULT_HISTORY_SUGGESTION_LIMIT
 ) : AwesomeBar.SuggestionProvider {
 
     override val id: String = UUID.randomUUID().toString()
@@ -47,20 +50,11 @@ class HistoryStorageSuggestionProvider(
             return emptyList()
         }
 
-        // Having both maxNumberOfResults and METADATA_SUGGESTION_LIMIT ensures that when asking for
-        // a lower number of suggestions the below filtering will have some that can be dropped.
-        val maxReturnedSuggestions = if (maxNumberOfResults > 0) {
-            minOf(maxNumberOfResults, HISTORY_SUGGESTION_LIMIT)
-        } else {
-            HISTORY_SUGGESTION_LIMIT
-        }
-
         // In case of duplicates we want to pick the suggestion with the highest score.
         // See: https://github.com/mozilla/application-services/issues/970
-        val suggestions = historyStorage.getSuggestions(text, HISTORY_SUGGESTION_LIMIT)
+        val suggestions = historyStorage.getSuggestions(text, maxNumberOfSuggestions)
             .sortedByDescending { it.score }
             .distinctBy { it.id }
-            .take(maxReturnedSuggestions)
 
         suggestions.firstOrNull()?.url?.let { url -> engine?.speculativeConnect(url) }