+++ mobile/android/android-components/components/feature/awesomebar/src/test/java/mozilla/components/feature/awesomebar/provider/HistoryMetadataSuggestionProviderTest.kt
@@ -16,7 +16,9 @@ import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Test
+import org.mockito.ArgumentMatchers.anyInt
 import org.mockito.ArgumentMatchers.anyString
+import org.mockito.Mockito.doReturn
 import org.mockito.Mockito.never
 import org.mockito.Mockito.times
 import org.mockito.Mockito.verify
@@ -43,7 +45,7 @@ class HistoryMetadataSuggestionProviderTest {
     fun `provider returns suggestions from configured history storage`() = runBlocking {
         val storage: HistoryMetadataStorage = mock()
 
-        whenever(storage.queryHistoryMetadata("moz", METADATA_SUGGESTION_LIMIT)).thenReturn(listOf(historyEntry))
+        whenever(storage.queryHistoryMetadata("moz", DEFAULT_METADATA_SUGGESTION_LIMIT)).thenReturn(listOf(historyEntry))
 
         val provider = HistoryMetadataSuggestionProvider(storage, mock())
         val suggestions = provider.onInputChanged("moz")
@@ -55,50 +57,41 @@ class HistoryMetadataSuggestionProviderTest {
     @Test
     fun `provider limits number of returned suggestions to 5 by default`() = runBlocking {
         val storage: HistoryMetadataStorage = mock()
-        val historyEntries = mutableListOf<HistoryMetadata>().apply {
-            for (i in 1..10) {
-                add(historyEntry)
-            }
-        }
-        whenever(storage.queryHistoryMetadata("moz", METADATA_SUGGESTION_LIMIT)).thenReturn(historyEntries)
+        doReturn(emptyList<HistoryMetadata>()).`when`(storage).queryHistoryMetadata(anyString(), anyInt())
         val provider = HistoryMetadataSuggestionProvider(storage, mock())
 
-        val suggestions = provider.onInputChanged("moz")
-        assertEquals(5, suggestions.size)
+        provider.onInputChanged("moz")
+
+        verify(storage).queryHistoryMetadata("moz", 5)
+        Unit
     }
 
     @Test
-    fun `provider allows lowering from outside the number of returned suggestions`() = runBlocking {
+    fun `provider allows lowering the number of returned suggestions beneath the default`() = runBlocking {
         val storage: HistoryMetadataStorage = mock()
-        val historyEntries = mutableListOf<HistoryMetadata>().apply {
-            for (i in 1..10) {
-                add(historyEntry)
-            }
-        }
-        whenever(storage.queryHistoryMetadata("moz", METADATA_SUGGESTION_LIMIT)).thenReturn(historyEntries)
+        doReturn(emptyList<HistoryMetadata>()).`when`(storage).queryHistoryMetadata(anyString(), anyInt())
         val provider = HistoryMetadataSuggestionProvider(
-            historyStorage = storage, loadUrlUseCase = mock(), maxNumberOfResults = 2
+            historyStorage = storage, loadUrlUseCase = mock(), maxNumberOfSuggestions = 2
         )
 
-        val suggestions = provider.onInputChanged("moz")
-        assertEquals(2, suggestions.size)
+        provider.onInputChanged("moz")
+
+        verify(storage).queryHistoryMetadata("moz", 2)
+        Unit
     }
 
     @Test
-    fun `provider doesn't allow increasing from outside the number of returned suggestions to past the default of 20`() = runBlocking {
+    fun `provider allows increasing the number of returned suggestions above the default`() = runBlocking {
         val storage: HistoryMetadataStorage = mock()
-        val historyEntries = mutableListOf<HistoryMetadata>().apply {
-            for (i in 1..10) {
-                add(historyEntry)
-            }
-        }
-        whenever(storage.queryHistoryMetadata("moz", METADATA_SUGGESTION_LIMIT)).thenReturn(historyEntries)
+        doReturn(emptyList<HistoryMetadata>()).`when`(storage).queryHistoryMetadata(anyString(), anyInt())
         val provider = HistoryMetadataSuggestionProvider(
-            historyStorage = storage, loadUrlUseCase = mock(), maxNumberOfResults = 8
+            historyStorage = storage, loadUrlUseCase = mock(), maxNumberOfSuggestions = 8
         )
 
-        val suggestions = provider.onInputChanged("moz")
-        assertEquals(5, suggestions.size)
+        provider.onInputChanged("moz")
+
+        verify(storage).queryHistoryMetadata("moz", 8)
+        Unit
     }
 
     @Test
@@ -108,7 +101,7 @@ class HistoryMetadataSuggestionProviderTest {
             add(historyEntry)
             add(historyEntry.copy(totalViewTime = 0))
         }
-        whenever(storage.queryHistoryMetadata("moz", METADATA_SUGGESTION_LIMIT)).thenReturn(historyEntries)
+        whenever(storage.queryHistoryMetadata("moz", DEFAULT_METADATA_SUGGESTION_LIMIT)).thenReturn(historyEntries)
         val provider = HistoryMetadataSuggestionProvider(storage, mock())
 
         val suggestions = provider.onInputChanged("moz")
@@ -131,7 +124,7 @@ class HistoryMetadataSuggestionProviderTest {
         assertTrue(suggestions.isEmpty())
         verify(engine, never()).speculativeConnect(anyString())
 
-        whenever(storage.queryHistoryMetadata("moz", METADATA_SUGGESTION_LIMIT)).thenReturn(listOf(historyEntry))
+        whenever(storage.queryHistoryMetadata("moz", DEFAULT_METADATA_SUGGESTION_LIMIT)).thenReturn(listOf(historyEntry))
 
         suggestions = provider.onInputChanged("moz")
         assertEquals(1, suggestions.size)